added mongo encryption
